supervisor/pyproject.toml
[project]
name = "validator_supervisor"
version = "1.0.0"
description = "Ethereum validator supervisor"
authors = [
    {name = "Jim Posen", email = "jim.posen@gmail.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"
requires-python = ">= 3.10"
dependencies = [
    "aiohttp ~=3.9.2",
		"marshmallow ~=3.20.2",
		"PyNaCl ~=1.4.0",
		"PyYAML ~=6.0",
		"python-prctl ~=1.8.1",
]

[project.optional-dependencies]
dev = [
    "mypy",
		"types-pyyaml",
		"types-setuptools",
		"pip-audit",
]

[project.urls]
Repository = "https://github.com/jimpo/eth-staking"

[build-system]
requires = ["uv_build >= 0.8.17, <0.9.0"]
build-backend = "uv_build"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
packages = "validator_supervisor"

[[tool.mypy.overrides]]
module = [
    "prctl",
		"nacl",
		"nacl.encoding",
		"nacl.exceptions",
		"nacl.hash",
		"nacl.pwhash",
		"nacl.secret",
		"nacl.utils",
		"nacl.setuptools",
]
ignore_missing_imports = true
